Fixed issue with deadbush name switch
