--- Assets/Scripts/Deadbush_Hurt.cs
@@ -1,16 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class DeadbushTrigger : MonoBehaviour
-{
-    private void OnTriggerEnter(Collider other)
-    {
-        HurtEffect hurtableObject = other.GetComponent<HurtEffect>();
-                
-        if (hurtableObject != null)
-        {
-          hurtableObject.Hit();
-        }
-    }
-}
