--- Assets/Scripts/Deadbush_Hurt.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 028e0b33c29b4994e9163534a27be418
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
